requirements.txt
netmiko
pylogger
requests
pyyaml
